+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Don't show disabled login options.
+- Add option to disable content warnings.
 
 ### Changed